+++ feature/compose/src/main/java/com/example/compose/NewScreen.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.ui.compose
+package com.example.compose
 
 import androidx.compose.runtime.Composable
