+++ feature/compose/src/main/java/com/example/compose/PlaygroundScreen.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.ui.compose
+package com.example.compose
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
@@ -13,10 +13,6 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.example.androidtechsample.R
-import com.example.androidtechsample.component.TextBlackHead1
-import com.example.androidtechsample.component.TextBlackHead2
-import com.example.androidtechsample.component.TextWhiteBody1
 
 @Composable
 fun PlaygroundScreen(navController: NavController) {
