+++ feature/compose/src/main/java/com/example/compose/Styles.kt
@@ -1,4 +1,4 @@
-package com.example.androidtechsample.resource
+package com.example.compose
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.res.colorResource
@@ -6,7 +6,6 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
-import com.example.androidtechsample.R
 
 @Composable
 fun textStyleBlack() = TextStyle(
