+++ feature/core/src/main/res/navigation/nav_main.xml
@@ -143,5 +143,5 @@
 
   <fragment
     android:id="@+id/fragment_compose"
-    android:name="com.example.androidtechsample.ui.compose.ComposeFragment" />
+    android:name="com.example.compose.ComposeFragment" />
 </navigation>
